# openapi.yaml
openapi: 3.0.0
info:
  title: REST API for Component Communication
  description: "REST APIs to Bridge Local Env"
  version: 1.0.0
paths:
  /health:
    get:
      summary: Health check
      responses:
        "200":
          description: Successfully running service

  /request/{user}/{component}/{path}:
    parameters:
      - name: user
        in: path
        required: true
        schema:
          type: string
      - name: component
        in: path
        required: true
        schema:
          type: string
      - name: path
        in: path
        required: true
        schema:
          type: string
    get:
      summary: Forward a GET request to the client
      responses:
        "200":
          description: Request forwarded successfully
    post:
      summary: Forward a POST request to the client
      responses:
        "200":
          description: Request forwarded successfully
    put:
      summary: Forward a PUT request to the client
      responses:
        "200":
          description: Request forwarded successfully
    delete:
      summary: Forward a DELETE request to the client
      responses:
        "200":
          description: Request forwarded successfully

  /request/{user}/{component}:
    parameters:
      - name: user
        in: path
        required: true
        schema:
          type: string
      - name: component
        in: path
        required: true
        schema:
          type: string
    get:
      summary: Forward a GET request to the client
      responses:
        "200":
          description: Request forwarded successfully
    post:
      summary: Forward a POST request to the client
      responses:
        "200":
          description: Request forwarded successfully
    put:
      summary: Forward a PUT request to the client
      responses:
        "200":
          description: Request forwarded successfully
    delete:
      summary: Forward a DELETE request to the client
      responses:
        "200":
          description: Request forwarded successfully

  /handle:
    post:
      summary: Handle request from client
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Request handled successfully